OAK-11090: Move creation of recnums/segrefs from segment data
When initializing a Segment instance, we create RecordNumbers/SegmentReferences instances with the data contained in SegmentData. This operation is currently done from two private methods, Segment#readRecordNumberOffsets and Segment#readReferencedSegments.
By moving this code somewhere else (e.g. in static methods of RecordNumbers/SegmentReferences), it makes it easier to reuse it and add unit tests for these methods.

This commit also add unit tests to verify these methods.

Closes OAK-11090
