+++ oak-segment-tar/src/main/java/org/apache/jackrabbit/oak/segment/SegmentReferences.java
@@ -17,10 +17,75 @@
 
 package org.apache.jackrabbit.oak.segment;
 
+import org.apache.jackrabbit.guava.common.collect.AbstractIterator;
+import org.apache.jackrabbit.oak.segment.data.SegmentData;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Iterator;
+
+import static org.apache.jackrabbit.guava.common.base.Preconditions.checkArgument;
+import static org.apache.jackrabbit.guava.common.base.Preconditions.checkState;
+
 /**
  * Represents a list of segment IDs referenced from a segment.
  */
 interface SegmentReferences extends Iterable<SegmentId> {
+
+    /** Builds a new instance of {@link SegmentReferences} from the provided {@link SegmentData}. */
+    static @NotNull SegmentReferences fromSegmentData(@NotNull SegmentData data, @NotNull SegmentIdProvider idProvider) {
+        final int referencedSegmentIdCount = data.getSegmentReferencesCount();
+
+        checkState(referencedSegmentIdCount + 1 < 0xffff, "Segment cannot have more than 0xffff references");
+
+        // We need to keep SegmentId references (as opposed to e.g. UUIDs)
+        // here as frequently resolving the segment ids via the segment id
+        // tables is prohibitively expensive.
+        // These SegmentId references are not a problem wrt. heap usage as
+        // their individual memoised references to their underlying segment
+        // is managed via the SegmentCache. It is the size of that cache that
+        // keeps overall heap usage by Segment instances bounded.
+        // See OAK-6106.
+
+        final SegmentId[] refIds = new SegmentId[referencedSegmentIdCount];
+
+        return new SegmentReferences() {
+
+            @Override
+            public SegmentId getSegmentId(int reference) {
+                checkArgument(reference <= referencedSegmentIdCount, "Segment reference out of bounds");
+                SegmentId id = refIds[reference - 1];
+                if (id == null) {
+                    synchronized (refIds) {
+                        id = refIds[reference - 1];
+                        if (id == null) {
+                            long msb = data.getSegmentReferenceMsb(reference - 1);
+                            long lsb = data.getSegmentReferenceLsb(reference - 1);
+                            id = idProvider.newSegmentId(msb, lsb);
+                            refIds[reference - 1] = id;
+                        }
+                    }
+                }
+                return id;
+            }
+
+            @NotNull
+            @Override
+            public Iterator<SegmentId> iterator() {
+                return new AbstractIterator<>() {
+                    private int reference = 1;
+
+                    @Override
+                    protected SegmentId computeNext() {
+                        if (reference <= referencedSegmentIdCount) {
+                            return getSegmentId(reference++);
+                        } else {
+                            return endOfData();
+                        }
+                    }
+                };
+            }
+        };
+    }
 
     /**
      * Fetch the referenced segment ID associated to the provided reference.
