+++ oak-segment-tar/src/test/java/org/apache/jackrabbit/oak/segment/SegmentReferencesTest.java
@@ -22,10 +22,14 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 
 import java.io.File;
+import java.io.IOException;
 
+import org.apache.jackrabbit.oak.commons.Buffer;
 import org.apache.jackrabbit.oak.plugins.memory.EmptyNodeState;
+import org.apache.jackrabbit.oak.segment.data.SegmentData;
 import org.apache.jackrabbit.oak.segment.file.FileStore;
 import org.apache.jackrabbit.oak.segment.file.FileStoreBuilder;
+import org.apache.jackrabbit.oak.segment.memory.MemoryStore;
 import org.apache.jackrabbit.oak.spi.state.NodeBuilder;
 import org.junit.Rule;
 import org.junit.Test;
@@ -92,4 +96,21 @@ public class SegmentReferencesTest {
         }
     }
 
+    @Test
+    public void shouldReadSegmentReferencesFromSegmentData() throws IOException {
+        SegmentData sampleSegmentData = SegmentData.newSegmentData(Buffer.wrap(new byte[] {
+                48, 97, 75, 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                -116, -71, -35, -50, 87, -7, 64, 125, -96, -24, -82, 112, 44, -36, 63, 67, 0, 0, 0, 0, 4, 0, 3, -1, -40,
+                0, 0, 0, 1, 5, 0, 3, -1, -48, 0, 0, 0, 2, 4, 0, 3, -1, -56, 0, 0, 0, 3, 5, 0, 3, -1, -64, 0, 0, 0, 0, 0,
+                0, 0, 0, 0, 0, 0, 0, 42, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 5, 0, 0, 10, -48, -66, -83, 11, -22, -48,
+                -66, 37, 123, 34, 119, 105, 100, 34, 58, 34, 116, 34, 44, 34, 115, 110, 111, 34, 58, 50, 44, 34, 116,
+                34, 58, 49, 55, 50, 51, 55, 49, 51, 51, 53, 57, 54, 54, 54, 125, 0, 0
+        }).asReadOnlyBuffer());
+
+        SegmentReferences segmentReferences = SegmentReferences.fromSegmentData(sampleSegmentData, new MemoryStore().getSegmentIdProvider());
+
+        SegmentId segmentId = segmentReferences.getSegmentId(1);
+        assertEquals(segmentId.getMostSignificantBits(), -8306364159399214979L);
+        assertEquals(segmentId.getLeastSignificantBits(), -6852035036232007869L);
+    }
 }
